refactor: vue directory structure
