+++ imake/vue/src/components/mode/Practice.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import SingleEffect from '@/components/shared/SingleEffect.vue'
+import SingleEffect from '@/components/templates/SingleEffect.vue'
 
 export default {
     name: 'Practice',
