+++ imake/vue/src/components/mode/custom/Custom.vue
@@ -50,10 +50,10 @@
 </template>
 
 <script>
-import Choices from '@/components/shared/Choices.vue'
+import Choices from '@/components/templates/Choices.vue'
 import CustomChoices from './CustomChoices.vue'
 export default {
-    name: 'CustomMode',
+    name: 'Custom',
     props: ['kwargs'],
     data: function () {
         return {
