+++ imake/vue/src/components/templates/Choices.vue
