+++ imake/vue/src/components/templates/SingleEffect.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import Choices from './Choices.vue'
+import Choices from '@/components/templates/Choices.vue'
 export default {
     name: 'SingleEffect',
     props: ['modeIconPath'],
